requirements.txt
aiodns==3.0.0
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
DateTime==4.3
frozenlist==1.2.0
future==0.18.2
idna==3.3
multidict==5.2.0
pyArango==1.3.5
pycares==4.1.2
pycparser==2.21
pytz==2021.3
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.2
zope.interface==5.4.0

SQLAlchemy~=1.4.26
python-dotenv~=0.19.1
GeoAlchemy2~=0.9.4
h3~=3.7.3